requirements.txt
aiocache~=0.12.0
pylibyaml~=0.1.0
PyYAML~=6.0